refactor(headless): isolate case assist bundle
https://coveord.atlassian.net/browse/KIT-2466
