+++ packages/headless/src/controllers/core/quickview/headless-core-quickview.ts
@@ -6,7 +6,6 @@ import {
 } from '../../../api/search/html/html-request';
 import {Result} from '../../../api/search/search/result';
 import {configuration} from '../../../app/common-reducers';
-import {resultPreview} from '../../../app/reducers';
 import {ClientThunkExtraArguments} from '../../../app/thunk-extra-arguments';
 import {
   fetchResultContent,
@@ -15,6 +14,7 @@ import {
   updateContentURL,
 } from '../../../features/result-preview/result-preview-actions';
 import {StateNeededByHtmlEndpoint} from '../../../features/result-preview/result-preview-request-builder';
+import {resultPreviewReducer as resultPreview} from '../../../features/result-preview/result-preview-slice';
 import {
   ConfigurationSection,
   ResultPreviewSection,
