+++ packages/headless/src/features/case-field/case-field-actions-loader.ts
@@ -1,7 +1,7 @@
 import {AsyncThunkAction, PayloadAction} from '@reduxjs/toolkit';
 import {AsyncThunkCaseAssistOptions} from '../../api/service/case-assist/case-assist-api-client';
 import {CaseAssistEngine} from '../../app/case-assist-engine/case-assist-engine';
-import {caseField} from '../../app/reducers';
+import {caseFieldReducer as caseField} from '../../features/case-field/case-field-slice';
 import {
   fetchCaseClassifications,
   FetchClassificationsThunkReturn,
