+++ packages/headless/src/features/document-suggestion/document-suggestion-actions-loader.ts
@@ -1,7 +1,7 @@
 import {AsyncThunkAction} from '@reduxjs/toolkit';
 import {AsyncThunkCaseAssistOptions} from '../../api/service/case-assist/case-assist-api-client';
 import {CaseAssistEngine} from '../../app/case-assist-engine/case-assist-engine';
-import {documentSuggestion} from '../../app/reducers';
+import {documentSuggestionReducer as documentSuggestion} from '../../features/document-suggestion/document-suggestion-slice';
 import {
   fetchDocumentSuggestions,
   FetchDocumentSuggestionsThunkReturn,
